+++ src/ofxArduinoDjShield.h
@@ -29,11 +29,11 @@ public:
 	float PotA1(bool rescale = false, int min = 0, int max = 100);
 	float PotA2(bool rescale = false, int min = 0, int max = 100);
 
-	bool ButtonD2(int DelayTime = 0);
-	bool ButtonD3(int DelayTime = 0);
-	bool ButtonD4(int DelayTime = 0);
-	bool ButtonD5(int DelayTime = 0);
-	bool ButtonD6(int DelayTime = 0);
+	bool ButtonD2(float DelayTime = 0);
+	bool ButtonD3(float DelayTime = 0);
+	bool ButtonD4(float DelayTime = 0);
+	bool ButtonD5(float DelayTime = 0);
+	bool ButtonD6(float DelayTime = 0);
 
 	void LedA4(bool Active);
 	void LedA5(bool Active);
